Remove dependencies, create simple client
