+++ lib/trexrb/version.rb
@@ -1,3 +1,3 @@
-module Trexrb
+class Trexrb
   VERSION = "0.1.1"
 end
